backend/requirements.txt
# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-multipart>=0.0.9
websockets>=12.0
aiofiles>=24.1.0

# Database
sqlalchemy[asyncio]>=2.0.35
aiosqlite>=0.20.0

# AI / ML
openai>=1.51.0
sentence-transformers>=3.1.1
chromadb>=0.5.5
numpy>=2.1.0

# Compliance
rank-bm25>=0.2.2

# Document Verification
python-Levenshtein>=0.25.1
jellyfish>=1.1.0

# Utilities
python-dotenv>=1.0.1
pydantic-settings>=2.5.2
httpx>=0.27.2

# Testing
pytest>=8.3.3
pytest-asyncio>=0.24.0
